Breaks apart more code. Adds a concept of progression / multiple levels. Starts users off with a simple tutorial level. No UI for the progression / high scores have been added yet, that will be a seperate PR after I think about it
