+++ .vscode/extensions.json
@@ -1,5 +1,6 @@
 {
     "recommendations": [
         "ritwickdey.liveserver",
+        "dbaeumer.vscode-eslint",
     ]
 }
